requirements.txt
flask
flask-cors
scikit-learn
pandas
numpy
joblib
python-docx
PyPDF2
pyyaml
streamlit
requests
Werkzeug
matplotlib
seaborn
catboost
xgboost
dill
pdfminer.six
flask
load_dotenv